requirements.txt
numpy
torch
torchvision
lightning

opencv-python
omegaconf
hydra-core
tensorboard

gymnasium
Box2D
swig
pygame
